Read of wheelchair tag for ways tagged areas=yes
The wheelchair tag of closed ways tagged "highway=pedestrian;area=yes" has been ignored so far. In case the area has the tag "wheelchair=no", the graph builder adds "wheelchair=yes" to the resulting graph. 

The reading of the wheelchair tag for those areas has been added in this commit.